veel toegevoegd, niets verbeterd

+++ be/kdg/rl/MDP.py
@@ -1,14 +1,14 @@
 import numpy
+from tabulate import tabulate
 
 from be.kdg.rl import Percept
 
 
 class MDP:
     def __init__(self, observation_space_size, action_space_size):
-        #TODO: Vragen aan Wouter:
         self._observation_space_size = observation_space_size
         self.action_space_size = action_space_size
-        self._Rtsa = numpy.zeros((observation_space_size, action_space_size, observation_space_size), dtype=int)
+        self._Rtsa = numpy.zeros((observation_space_size, action_space_size, observation_space_size))
         self._Nsa = numpy.zeros((observation_space_size, action_space_size), dtype=int)
         self._Ntsa = numpy.zeros((observation_space_size, action_space_size, observation_space_size), dtype=int)
         self._Ptsa = numpy.zeros((observation_space_size, action_space_size, observation_space_size))
